refactor(Core UI): :recycle: Modular css
