+++ src/App.css
@@ -3,171 +3,10 @@ html {
   height: 100%;
 }
 
-/* Layout */
 #app {
   height: 100%;
 }
 
-.pf-v5-c-nav__link {
-  transition: 0.3s;
-}
-
-.pf-v5-c-page__sidebar.pf-m-collapsed {
-  transition: max-width 0.2s ease-in-out;
-}
-
-.pf-v5-c-page__sidebar.pf-m-expanded {
-  transition: max-width 0.2s ease-in-out;
-  max-width: 100%;
-}
-
-/* Resource icon */
-.sk-resource-icon {
-  background-color: var(--pf-v5-global--palette--black-500);
-  border-radius: 20px;
-  color: var(--pf-v5-global--palette--white);
-  display: inline-block;
-  flex-shrink: 0;
-  font-size: 12px;
-  font-weight: 400;
-  line-height: 20px;
-  margin-right: 4px;
-  min-width: 20px;
-  padding: 0 4px;
-  text-align: center;
-  white-space: nowrap;
-}
-
-.sk-resource-site {
-  background-color: var(--pf-v5-global--palette--green-600);
-}
-
-.sk-resource-process-group {
-  background-color: var(--pf-v5-global--palette--light-green-500);
-}
-
-.sk-resource-service {
-  background-color: var(--pf-v5-global--palette--purple-500);
-}
-
-.sk-resource-process {
-  background-color: var(--pf-v5-global--palette--cyan-300);
-}
-
-.sk-resource-skupper {
-  margin-left: -5px;
-  min-width: 32px;
-  background-color: transparent;
-}
-
-/* Loader */
-.sk-loading-page .cog-wrapper {
-  position: relative;
-  margin: auto;
-  width: 5.5em;
-  height: 5em;
-  color: var(--pf-v5-global--palette--black-500);
-}
-
-.sk-loading-page .cog-wrapper .cog {
-  position: absolute;
-}
-
-.sk-loading-page .cog-wrapper .cog-main {
-  width: 4em;
-  height: 4em;
-  top: 0.5em;
-  left: 0;
-}
-
-.sk-loading-page .cog-wrapper .cog-secondary {
-  width: 2em;
-  height: 2em;
-}
-
-.sk-loading-page .cog-wrapper .cog-upper {
-  top: 0;
-  right: 0;
-}
-
-.sk-loading-page .cog-wrapper .cog-lower {
-  bottom: 0;
-  right: 0;
-}
-
-.sk-loading-page .cog-wrapper .spinning-clockwise {
-  animation: spin 4s linear infinite;
-}
-
-.sk-loading-page .cog-wrapper .spinning-clockwise--reverse {
-  animation: spin 4s linear infinite reverse;
-}
-
-@keyframes spin {
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-.sk-topology-controls {
-  position: absolute !important;
-  bottom: 0;
-  left: 0;
-  background-color: transparent !important;
-}
-
-.sk-topology-control-bar__button {
-  margin-right: var(--pf-v5-global--spacer--xs);
-  margin-top: var(--pf-v5-global--spacer--xs);
-  border: none;
-  border-radius: var(--pf-v5-global--BorderRadius--sm);
-  box-shadow: var(--pf-v5-global--BoxShadow--sm);
-}
-.sk-topology-control-bar__button:not(.pf-v5-m-disabled) {
-  background-color: var(--pf-v5-global--BackgroundColor--100);
-}
-.sk-topology-control-bar__button:after {
-  display: none;
-}
-.sk-topology-control-bar__button:hover {
-  border: none;
-  box-shadow: var(--pf-v5-global--BoxShadow--md);
-}
-
 .sk-capitalize {
   text-transform: capitalize;
-}
-
-.metric-disabled {
-  --pf-c-card--BackgroundColor: var(--pf-v5-global--palette--black-200) !important;
-  color: var(--pf-v5-global--palette--black-500);
-}
-
-.sk-search-filter input {
-  padding-left: 8px;
-}
-
-.sk-search-filter .pf-v5-c-text-input-group__icon {
-  display: none;
-}
-
-/* refresh metrics button */
-.button-toggle-dropdown-loading button {
-  background-color: var(--pf-v5-global--danger-color--100) !important;
-}
-
-.button-toggle-loading {
-  background-color: var(--pf-v5-global--danger-color--100) !important;
-}
-
-.button-toggle-off {
-  background-color: var(--pf-v5-global--palette-black--300) !important;
-}
-
-.button-toggle-spinner .pf-v5-c-spinner__path {
-  stroke: var(--pf-v5-global--palette--white);
-}
-
-.button-toggle-spinner-text {
-  margin-left: var(--pf-v5-global--spacer--sm);
 }
