+++ src/core/components/SkUpdateDataButton/SkUpdateDataButton.spec.tsx
@@ -4,10 +4,9 @@ import { Server } from 'miragejs';
 import { Wrapper } from '@core/components/Wrapper';
 import { loadMockServer } from '@mocks/server';
 
-import UpdateMetricsButton from '../components/UpdateMetricsButton';
-import { refreshDataIntervalMap } from '../Metrics.constants';
+import SkUpdateDataButton, { refreshDataIntervalMap } from '.';
 
-describe('Metrics component', () => {
+describe('SkUpdateDataButton component', () => {
   let server: Server;
   beforeEach(() => {
     server = loadMockServer() as Server;
@@ -19,13 +18,13 @@ describe('Metrics component', () => {
     jest.clearAllMocks();
   });
 
-  it('should render the Metric Filter', async () => {
+  it('should render the Update Data Button and simulate a click', async () => {
     const onClickMock = jest.fn();
     const onRefreshIntervalSelectedMock = jest.fn();
 
     render(
       <Wrapper>
-        <UpdateMetricsButton
+        <SkUpdateDataButton
           refreshIntervalDefault={refreshDataIntervalMap[1].value}
           isLoading={false}
           onClick={onClickMock}
@@ -34,10 +33,10 @@ describe('Metrics component', () => {
       </Wrapper>
     );
 
-    fireEvent.click(screen.getByTestId('update-metric-click'));
+    fireEvent.click(screen.getByTestId('update-data-click'));
     expect(onClickMock).toBeCalledTimes(1);
 
-    fireEvent.click(screen.getByTestId('update-metric-dropdown'));
+    fireEvent.click(screen.getByTestId('update-data-dropdown'));
     // use waitFor to avoid this  Warning: An update to Popper inside a test was not wrapped in act(...).
     await waitFor(() => expect(screen.getByText(refreshDataIntervalMap[2].label)).toBeInTheDocument);
