+++ src/layout/SideBar.tsx
@@ -2,6 +2,8 @@ import { PageSidebar, PageSidebarBody } from '@patternfly/react-core';
 
 import NavBar from '@core/components/NavBar';
 
+import './Sidebar.css';
+
 const SkSidebar = function () {
   return (
     <PageSidebar>
