+++ src/pages/shared/Loading/Loading.css
@@ -1,0 +1,47 @@
+.sk-loading-page .cog-wrapper {
+  position: relative;
+  margin: auto;
+  width: 5.5em;
+  height: 5em;
+  color: var(--pf-v5-global--palette--black-500);
+}
+
+.sk-loading-page .cog-wrapper .cog {
+  position: absolute;
+}
+
+.sk-loading-page .cog-wrapper .cog-main {
+  width: 4em;
+  height: 4em;
+  top: 0.5em;
+  left: 0;
+}
+
+.sk-loading-page .cog-wrapper .cog-secondary {
+  width: 2em;
+  height: 2em;
+}
+
+.sk-loading-page .cog-wrapper .cog-upper {
+  top: 0;
+  right: 0;
+}
+
+.sk-loading-page .cog-wrapper .cog-lower {
+  bottom: 0;
+  right: 0;
+}
+
+.sk-loading-page .cog-wrapper .spinning-clockwise {
+  animation: spin 4s linear infinite;
+}
+
+.sk-loading-page .cog-wrapper .spinning-clockwise--reverse {
+  animation: spin 4s linear infinite reverse;
+}
+
+@keyframes spin {
+  100% {
+    transform: rotate(360deg);
+  }
+}
