+++ src/pages/shared/Metrics/Metrics.constants.ts
@@ -1,34 +1,6 @@
 import { MetricsLabels } from './Metrics.enum';
 import { ConfigMetricFilters } from './Metrics.interfaces';
 
-export const refreshDataIntervalMap = [
-  {
-    key: 'pause',
-    value: 0,
-    label: 'Off'
-  },
-  {
-    key: '20s',
-    value: 20 * 1000,
-    label: '20s'
-  },
-  {
-    key: '40s',
-    value: 40 * 1000,
-    label: '40s'
-  },
-  {
-    key: '60s',
-    value: 60 * 1000,
-    label: '1m'
-  },
-  {
-    key: '120s',
-    value: 120 * 1000,
-    label: '2m'
-  }
-];
-
 // default values to enable/disable filters
 export const configDefaultFilters: ConfigMetricFilters = {
   sourceSites: { disabled: false, placeholder: MetricsLabels.FilterAllSourceSites, hide: false },
